# product-service/docs/swagger.yaml
definitions:
...
  request.ProductRequest:
    properties:
      category_slug:
        maxLength: 100
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
      status:
        type: string
      unit:
        maxLength: 100
        type: string
      variant:
        type: integer
      variant_detail:
        items:
          $ref: '#/definitions/request.ProductDetailRequest'
        type: array
    required:
    - category_slug
    - description
    - name
    - status
    - unit
    - variant
    - variant_detail
    type: object
...
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  response.DefaultResponseWithPagination:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
...
  response.ProductDetailResponse:
    properties:
      category_name:
        type: string
      category_slug:
        type: string
      child:
        items:
          $ref: '#/definitions/response.ProductChildResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      parent_id:
        type: integer
      reguler_price:
        type: integer
      sale_price:
        type: integer
      status:
        type: string
      stock:
        type: integer
      unit:
        type: string
      weight:
        type: integer
    type: object
...
  response.ProductHomeListResponse:
    properties:
      category_name:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      reguler_price:
        type: integer
      sale_price:
        type: integer
    type: object
  response.ProductListResponse:
    properties:
      category_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      reguler_price:
        type: integer
      sale_price:
        type: integer
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/products:
    get:
      consumes:
      - application/json
      description: Get paginated list of products with filters for admin
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Order by (price_asc, price_desc, newest)
        in: query
        name: orderBy
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Filter by minimum price
        in: query
        name: startPrice
        type: integer
      - description: Filter by maximum price
        in: query
        name: endPrice
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by category slug
        in: query
        name: category
        type: string
      - description: Filter by parent product
        in: query
        name: is_parent
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponseWithPagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.ProductListResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get all products (admin)
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with its variants
      parameters:
      - description: Create Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Category Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create product (admin)
      tags:
      - products
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete product (admin)
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get product detail by ID for admin
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ProductDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get product by ID (admin)
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update product (admin)
      tags:
      - products
...
  /products:
    get:
      consumes:
      - application/json
      description: Get paginated list of active products with filters for the shop
        page
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Filter by category slug
        in: query
        name: category
        type: string
      - description: Order by (price_asc, price_desc, newest)
        in: query
        name: orderBy
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Filter by minimum price
        in: query
        name: startPrice
        type: integer
      - description: Filter by maximum price
        in: query
        name: endPrice
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponseWithPagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.ProductListResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get all shop products (public)
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information of a specific product by ID for the home/shop
        page
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ProductHomeDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get product detail (public)
      tags:
      - products
  /products/featured:
    get:
      consumes:
      - application/json
      description: Get a limited list of featured products for the home page
      parameters:
      - description: 'Items limit (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.ProductHomeListResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get featured products (public)
      tags:
      - products
swagger: "2.0"